Phase C: extract pure csvimporter + qifimporter modules
Split CSV/QIF import logic out of ui/core and db/core into two new db-free
modules that depend only on importengineapi + model/core (+ utils), enforced by
the pure-importer-convention's verifyNoDbDependency guard.

- app/csvimporter: CsvTransferMapper, StrategyMatcher, CsvImportApplier,
  the pure account-selection helpers (extracted from ApplyStrategyDialog),
  BulkImportResult, CsvImportResult. Package com.moneymanager.csvimporter.
- app/qifimporter (depends on csvimporter): QifCsvAdapter, QifImportApplier,
  QifImportResult. Package com.moneymanager.qifimporter.
- Relocate StringSimilarity into importengineapi (so csvimporter can use it
  without depending on the engine impl).
- Hoist shared constants into model/core: WellKnownIds (well-known attr/
  relationship type ids) and QifColumns (fixed QIF column names), so the db
  seeding and the importers reference one source of truth without db coupling.
- Pure unit tests move into the new modules; db-integration tests stay in
  db/core with testImplementation(csvimporter, qifimporter).
- Rewire ui/core consumers + build files; buildHealth/detekt/ktlint clean.
