+++ app/csvimporter/src/commonMain/kotlin/com/moneymanager/csvimporter/StrategyMatcher.kt
@@ -1,4 +1,4 @@
-package com.moneymanager.database.csv
+package com.moneymanager.csvimporter
 
 import com.moneymanager.domain.model.csvstrategy.CsvImportStrategy
