+++ app/csvimporter/src/commonTest/kotlin/com/moneymanager/csvimporter/CsvAccountMappingTest.kt
@@ -1,6 +1,6 @@
 @file:OptIn(kotlin.time.ExperimentalTime::class, kotlin.uuid.ExperimentalUuidApi::class)
 
-package com.moneymanager.database.csv
+package com.moneymanager.csvimporter
 
 import com.moneymanager.domain.model.Account
 import com.moneymanager.domain.model.AccountId
