+++ app/csvimporter/src/commonTest/kotlin/com/moneymanager/csvimporter/CsvDuplicateDetectionTest.kt
@@ -1,6 +1,6 @@
 @file:OptIn(kotlin.time.ExperimentalTime::class, kotlin.uuid.ExperimentalUuidApi::class)
 
-package com.moneymanager.database.csv
+package com.moneymanager.csvimporter
 
 import com.moneymanager.bigdecimal.BigDecimal
 import com.moneymanager.domain.model.AccountId
