+++ app/db/core/build.gradle.kts
@@ -28,8 +28,10 @@ kotlin {
             dependencies {
                 implementation(kotlin("test"))
                 implementation(libs.kotlinx.coroutines.test)
+                implementation(projects.app.csvimporter)
                 implementation(projects.app.importengineapi)
                 implementation(projects.app.model.core)
+                implementation(projects.app.qifimporter)
                 implementation(projects.test.app.db)
             }
         }
