+++ app/db/core/src/commonTest/kotlin/com/moneymanager/database/csv/CsvImportDuplicateDetectionIntegrationTest.kt
@@ -2,13 +2,15 @@
 
 package com.moneymanager.database.csv
 import com.moneymanager.bigdecimal.BigDecimal
+import com.moneymanager.csvimporter.CsvTransferMapper
+import com.moneymanager.csvimporter.ExistingTransferInfo
 import com.moneymanager.domain.model.Account
 import com.moneymanager.domain.model.AccountId
 import com.moneymanager.domain.model.Currency
 import com.moneymanager.domain.model.DeviceInfo
 import com.moneymanager.domain.model.Money
-import com.moneymanager.domain.model.Source
 import com.moneymanager.domain.model.NewAttribute
+import com.moneymanager.domain.model.Source
 import com.moneymanager.domain.model.Transfer
 import com.moneymanager.domain.model.TransferId
 import com.moneymanager.domain.model.csv.CsvColumn
