+++ app/db/core/src/commonTest/kotlin/com/moneymanager/database/csv/WiseCsvMapperTest.kt
@@ -2,6 +2,10 @@
 
 package com.moneymanager.database.csv
 
+import com.moneymanager.csvimporter.CsvTransferMapper
+import com.moneymanager.csvimporter.ExistingTransferInfo
+import com.moneymanager.csvimporter.MappingResult
+import com.moneymanager.csvimporter.NewAccount
 import com.moneymanager.database.DatabaseConfig
 import com.moneymanager.domain.model.Account
 import com.moneymanager.domain.model.AccountId
