+++ app/ui/core/build.gradle.kts
@@ -11,9 +11,11 @@ kotlin {
                 api(libs.kotlinx.coroutines.core)
                 api(libs.kotlinx.datetime)
                 api(libs.kotlinx.serialization.json)
+                api(projects.app.csvimporter)
                 api(projects.app.db.core)
                 api(projects.app.importer)
                 api(projects.app.importengineapi)
+                api(projects.app.qifimporter)
                 api(projects.app.model.core)
                 api(projects.utils.bigdecimal)
                 api(projects.utils.rest)
@@ -74,9 +76,11 @@ kotlin {
                 api(libs.compose.ui.desktop)
                 api(libs.compose.ui.geometry.desktop)
                 api(libs.compose.ui.unit.desktop)
+                api(projects.app.csvimporter)
                 api(projects.app.db.core)
                 api(projects.app.importengineapi)
                 api(projects.app.model.core)
+                api(projects.app.qifimporter)
                 api(projects.utils.bigdecimal)
                 api(projects.utils.rest)
