+++ app/ui/core/src/commonMain/kotlin/com/moneymanager/ui/screens/csv/CsvImportAllDialog.kt
@@ -16,7 +16,9 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.moneymanager.database.csv.StrategyMatcher
+import com.moneymanager.csvimporter.StrategyMatcher
+import com.moneymanager.csvimporter.bulkApplyCsv
+import com.moneymanager.csvimporter.needsSourceAccountOverride
 import com.moneymanager.domain.Maintenance
 import com.moneymanager.domain.model.AccountId
 import com.moneymanager.domain.model.csv.CsvImport
