+++ app/ui/core/src/commonMain/kotlin/com/moneymanager/ui/screens/qif/QifStrategyEditorScreen.kt
@@ -11,7 +11,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import com.moneymanager.database.qif.QifCsvAdapter
 import com.moneymanager.domain.model.csv.CsvRow
 import com.moneymanager.domain.model.csvstrategy.CsvImportStrategyId
 import com.moneymanager.domain.model.qif.QifImportId
@@ -25,6 +24,7 @@ import com.moneymanager.domain.repository.CurrencyRepository
 import com.moneymanager.domain.repository.PersonAccountOwnershipRepository
 import com.moneymanager.domain.repository.PersonRepository
 import com.moneymanager.domain.repository.QifImportRepository
+import com.moneymanager.qifimporter.QifCsvAdapter
 import com.moneymanager.ui.screens.csvstrategy.editor.CsvStrategyEditorScreen
 
 /**
